@@ -4,18 +4,18 @@ description: Interface para sistemas de integração
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.2
-  http: ^0.12.0+4
-  universal_html: ^1.2.3
-  flutter_dotenv: ^2.1.0
-  url_launcher: ^5.4.2
-  flutter_typeahead: ^1.8.8
-  auto_size_text: ^2.1.0
+  cupertino_icons: ^1.0.4
+  http: ^0.13.4
+  universal_html: ^2.0.8
+  flutter_dotenv: ^5.0.2
+  url_launcher: ^6.0.20
+  flutter_typeahead: ^3.2.4
+  auto_size_text: ^3.0.0
 
 dev_dependencies:
   flutter_test:
