@@ -11,7 +11,7 @@ class Message {
   final double width;
   final double height;
 
-  Widget componentWidget;
+  Widget? componentWidget;
   Map<String, dynamic> componentConfigs = {
     "protocol": [
       {
@@ -40,7 +40,7 @@ class Message {
   };
 
   Map<String, dynamic> updateConfigs(selectedItem, config, configControllers) {
-    var protocolOptions = List();
+    var protocolOptions = <String>[];
 
     if (configControllers["editItemsValues"][config] != null) {
       for (var protocolOption in selectedItem.componentConfigs[config][0]
@@ -60,7 +60,7 @@ class Message {
 
   List<Widget> buildEditPane(selectedItem, selectedItemID, itemsPositions,
       config, configControllers, editItems, baseWidget) {
-    String _chosenValue;
+    String? _chosenValue;
 
     if (selectedItem.componentConfigs[config][1] != null)
       _chosenValue = selectedItem.componentConfigs[config][1];
@@ -83,7 +83,7 @@ class Message {
             icon: Icon(Icons.arrow_downward),
             iconSize: 24,
             elevation: 16,
-            onChanged: (String newValue) {
+            onChanged: (String? newValue) {
               baseWidget.setState(() {
                 configControllers["editItemsValues"][config] = newValue;
                 baseWidget.componentConfigControllers["editItemsValues"]
@@ -112,7 +112,7 @@ class Message {
         configControllers["messageControllers"]
             .addAll({protocolOption: TextEditingController()});
 
-        String controllerTextValue = "";
+        String? controllerTextValue = "";
         if (protocolOptionIndex <
             selectedItem.componentConfigs[config][2].length) {
           controllerTextValue =
@@ -174,7 +174,7 @@ class Message {
     return SizedBox(
         width: 100,
         child: Draggable(
-          feedback: componentWidget,
+          feedback: componentWidget!,
           onDraggableCanceled: (velocity, offset) {
             insertNewItem(this, offset);
           },
